requirements.txt
python-telegram-bot[job-queue]==21.6
anthropic==0.34.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
apscheduler==3.10.4
python-dotenv==1.0.1
aiofiles==24.1.0
